+++ integration_test/player_backend_stability_matrix_test.dart
@@ -7,6 +7,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:local_tag_player/src/app.dart';
+import 'package:local_tag_player/src/features/player/application/player_fullscreen_lifecycle_controller.dart';
 import 'package:media_kit/media_kit.dart';
 
 // ignore_for_file: slash_for_doc_comments